moodboard

+++ Portfolio-Api/Services/Repositories/Comment/CommentRepository.cs
@@ -1,5 +1,7 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using Google.Apis.Auth;
+using Microsoft.EntityFrameworkCore;
 using Portfolio_Api.Models;
+using Portfolio_Api.Models.RequestModels;
 
 namespace Portfolio_Api.Services.Repositories
 {
@@ -15,23 +17,45 @@ namespace Portfolio_Api.Services.Repositories
             _context = context;
         }
 
-        public async Task AddCommentAsync(Comment comment)
+        public async Task AddCommentAsync(CommentRequest commentRequest)
         {
-            _context.comments.Add(comment);
+            var validatedComment = new Comment()
+                .setComment(commentRequest.comment);
+            _context.comments.Add(validatedComment);
+            await _context.SaveChangesAsync();
+        }
+        public async Task AddCommentAsync(CommentRequest commentRequest, string jwt)
+        {
+            GoogleJsonWebSignature.ValidationSettings settings = new GoogleJsonWebSignature.ValidationSettings();
+            settings.Audience = new List<string>() { configuration.GetSection("AppSettings:GoogleId").Value };
+            GoogleJsonWebSignature.Payload payload = GoogleJsonWebSignature.ValidateAsync(jwt, settings).Result;
+            var validatedComment = new Comment()
+                .setComment(commentRequest.comment)
+                .setPosterUsername(payload.Name);
+            _context.comments.Add(validatedComment);
             await _context.SaveChangesAsync();
         }
 
-        public async Task DeleteCommentAsync(Comment comment)
+        public async Task DeleteCommentAsync(CommentRequest comment, string jwt)
         {
+            GoogleJsonWebSignature.ValidationSettings settings = new GoogleJsonWebSignature.ValidationSettings();
+            settings.Audience = new List<string>() { configuration.GetSection("AppSettings:GoogleId").Value };
+            GoogleJsonWebSignature.Payload jwtPayload = GoogleJsonWebSignature.ValidateAsync(jwt, settings).Result;
+
             var commentToDelete = await _context.comments.FirstOrDefaultAsync(_comment => _comment.id == comment.id);
-            if (commentToDelete != null)
+
+            if (commentToDelete != null && commentToDelete.poster_username == jwtPayload.Name)
             {
                 _context.comments.Remove(commentToDelete);
                 await _context.SaveChangesAsync();
             }
+            else
+            {
+                throw new UnauthorizedAccessException();
+            }
         }
 
-        public async Task EditCommentAsync(Comment comment)
+        public async Task EditCommentAsync(CommentRequest comment)
         {
             var commentToEdit = await _context.comments.FirstAsync(_comment => _comment.id == comment.id);
             if (commentToEdit != null)
@@ -46,14 +70,9 @@ namespace Portfolio_Api.Services.Repositories
             return await _context.comments.FirstAsync(comment => comment.id == id);
         }
 
-        public IEnumerable<Comment> GetBlogPosts()
+        public async Task<IEnumerable<Comment>> GetCommentsAsync()
         {
-            return _context.comments.ToList();
-        }
-
-        public IEnumerable<Comment> GetComments()
-        {
-            throw new NotImplementedException();
+            return await _context.comments.ToListAsync();
         }
     }
 }
